requirements.txt
# FastAPI 和 Web 框架
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# 数据处理
pandas==2.1.3
openpyxl==3.11.0
numpy==1.24.3

# 机器学习
scikit-learn==1.3.2
joblib==1.3.2

# 可视化
matplotlib==3.8.2

# 火山引擎大模型 SDK
volcengine-python-sdk==1.0.12

# 环境变量管理
python-dotenv==1.0.0

# 其他工具
requests==2.31.0


pydantic